refactor(ingest/hex): extract HexSourceConfig to a separate file and update imports
