+++ metadata-ingestion/src/datahub/ingestion/source/hex/config.py
@@ -1,0 +1,129 @@
+from typing import Dict, Optional
+
+from pydantic import Field, SecretStr
+
+from datahub.configuration.common import AllowDenyPattern
+from datahub.configuration.source_common import (
+    EnvConfigMixin,
+    PlatformInstanceConfigMixin,
+)
+from datahub.configuration.validate_field_removal import pydantic_removed_field
+from datahub.ingestion.source.hex.constants import (
+    HEX_API_BASE_URL_DEFAULT,
+    HEX_API_PAGE_SIZE_DEFAULT,
+)
+from datahub.ingestion.source.state.stale_entity_removal_handler import (
+    StatefulStaleMetadataRemovalConfig,
+)
+from datahub.ingestion.source.state.stateful_ingestion_base import (
+    StatefulIngestionConfigBase,
+)
+
+
+class HexSourceConfig(
+    StatefulIngestionConfigBase, PlatformInstanceConfigMixin, EnvConfigMixin
+):
+    workspace_name: str = Field(
+        description="Hex workspace name. You can find this name in your Hex home page URL: https://app.hex.tech/<workspace_name>",
+    )
+    token: SecretStr = Field(
+        description=(
+            "Hex Workspace Token with the 'Read projects' scope. "
+            "Create one at Settings → API → Workspace tokens. "
+            "The 'Read projects' scope is required to access project cells for lineage; "
+            "tokens without it can enumerate projects but not read their content. "
+            "See https://learn.hex.tech/docs/api-integrations/api/overview for token types."
+        ),
+    )
+    base_url: str = Field(
+        default=HEX_API_BASE_URL_DEFAULT,
+        description="Hex API base URL. For most Hex users, this will be https://app.hex.tech/api/v1. "
+        "Single-tenant app users should replace this with the URL they use to access Hex.",
+    )
+    stateful_ingestion: Optional[StatefulStaleMetadataRemovalConfig] = Field(
+        default=None,
+        description="Configuration for stateful ingestion and stale metadata removal.",
+    )
+    include_components: bool = Field(
+        default=True,
+        description="Include Hex Components in the ingestion",
+    )
+    page_size: int = Field(
+        default=HEX_API_PAGE_SIZE_DEFAULT,
+        description="Number of items to fetch per Hex API call.",
+    )
+    patch_metadata: bool = Field(
+        default=False,
+        description="Emit metadata as patch events",
+    )
+    collections_as_tags: bool = Field(
+        default=True,
+        description="Emit Hex Collections as tags",
+    )
+    status_as_tag: bool = Field(
+        default=True,
+        description="Emit Hex Status as tags",
+    )
+    categories_as_tags: bool = Field(
+        default=True,
+        description="Emit Hex Category as tags",
+    )
+    project_title_pattern: AllowDenyPattern = Field(
+        default=AllowDenyPattern.allow_all(),
+        description="Regex pattern for project titles to filter in ingestion.",
+    )
+    component_title_pattern: AllowDenyPattern = Field(
+        default=AllowDenyPattern.allow_all(),
+        description="Regex pattern for component titles to filter in ingestion.",
+    )
+    set_ownership_from_email: bool = Field(
+        default=True,
+        description="Set ownership identity from owner/creator email",
+    )
+    include_lineage: bool = Field(
+        default=True,
+        description="Extract upstream lineage. Uses queriedTables API (ENTERPRISE workspaces) "
+        "or falls back to parsing SQL from cells (all workspaces). No warehouse ingestion "
+        "dependency required.",
+    )
+    connection_platform_map: Dict[str, str] = Field(
+        default_factory=dict,
+        description="Map connection ID to DataHub platform name for connections that cannot "
+        "be resolved automatically (deleted connections, permission gaps, unsupported types). "
+        "Lineage is skipped — not guessed — for unmapped connections. "
+        'Example: {"<connection_uuid>": "snowflake"}',
+    )
+    include_run_history: bool = Field(
+        default=True,
+        description="Emit the most recent scheduled run as an Operation aspect.",
+    )
+    include_context_documents: bool = Field(
+        default=True,
+        description="Emit a DataHub Document per project containing SQL sources, "
+        "visualisation metadata, and notebook documentation. Documents are hidden from "
+        "global search and linked to the project Dashboard for AI agent retrieval.",
+    )
+    category_pattern: AllowDenyPattern = Field(
+        default=AllowDenyPattern.allow_all(),
+        description="Regex pattern for categories to filter in ingestion. This will exclude any project or component that has any category denied or not explicitly allowed.",
+    )
+    max_projects: Optional[int] = Field(
+        default=None,
+        description="Maximum number of projects to process. Useful for testing or staged rollouts. "
+        "Components discovered during project processing are not counted. "
+        "Defaults to None (process all projects).",
+    )
+
+    # Removed fields — emit a clear warning rather than silently ignoring.
+    # These were used by the old DataHub-query-fetcher lineage path which searched
+    # DataHub for Query entities tagged with Hex metadata comments. Lineage now comes
+    # directly from the Hex REST API and none of these fields have any effect.
+    _lineage_start_time_removed = pydantic_removed_field(
+        "lineage_start_time", month="May", year=2026
+    )
+    _lineage_end_time_removed = pydantic_removed_field(
+        "lineage_end_time", month="May", year=2026
+    )
+    _datahub_page_size_removed = pydantic_removed_field(
+        "datahub_page_size", month="May", year=2026
+    )
